requirements.txt
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.4.2
scikit-image==0.24.0
opencv-python==4.8.1.78
matplotlib==3.8.4
pypdf2==3.0.1
Pillow==10.3.0
tqdm==4.67.1
reportlab==4.2.0